+++ scrapli_community/raisecom/ros/sync_driver.py
@@ -1,0 +1,39 @@
+"""scrapli_community.raisecom.ros.sync_driver"""
+from scrapli.driver import NetworkDriver
+
+
+def default_sync_on_open(conn: NetworkDriver) -> None:
+    """
+    raisecom_ros on_open callable
+
+    Args:
+        conn: NetworkDriver object
+
+    Returns:
+        N/A
+
+    Raises:
+        N/A
+    """
+    conn.acquire_priv(desired_priv=conn.default_desired_privilege_level)
+    conn.send_command(command="terminal page-break disable")
+
+
+def default_sync_on_close(conn: NetworkDriver) -> None:
+    """
+    raisecom_ros on_close callable
+
+    Args:
+        conn: NetworkDriver object
+
+    Returns:
+        N/A
+
+    Raises:
+        N/A
+    """
+    # write exit directly to the transport as channel would fail to find the prompt after sending
+    # the exit command!
+    conn.acquire_priv(desired_priv=conn.default_desired_privilege_level)
+    conn.channel.write(channel_input="exit")
+    conn.channel.send_return()