+++ scrapli_community/__init__.py
@@ -1,3 +1,3 @@
 """scrapli_community"""
 
-__version__ = "2022.07.30"
+__version__ = "2023.01.30"