+++ docs/changelog.md
@@ -4,6 +4,7 @@ CHANGELOG
 # 2023.01.30 _(in development)_
 - Add Ruckus Unleashed -- see #143
 - Add Cumulus Linux & vtysh -- see #144
+- Add Raisecom ROS (thank you to @ikievite)
 
 
 # 2022.07.30